Remove all old Bokeh plotting code from webapp
The plots that were originally shown in the UI have been unused in the webapp since we switched to the tables interface some time ago.

This led to there being a *lot* of dead code, some of which was completely unused and some of which was actually still doing stuff in the background but without any visible results, basically just wasting CPU time.

This commit:
 - removes the layout module that was used to position objects on the plots
 - updates the webapp HTML template to remove various inactive/unused widgets
 - strips all the plot-specific code from server_lifecycle.py
 - does the same for main.py
 - updates the handlers for the table-based scoring to get rid of some hacky workarounds linked to the existence of the plot-based scoring code
 - fixes up some linter warnings on unused variables and imports