Update linker database schema
Previously most columns were treated as TEXT, which is fine for storing
most types of data involved here. However it breaks sorting in the bokeh
DataTables because it only sorts numerically if it detects the source
data is not text. So this commit adjusts the datatype for each column of
each table such that only the ones that actually store text will have
the TEXT datatype.