Bugfix for "include_strains" feature
When the user provides an include_strains.csv file this indicates that
only the strains listed in it be retained, any objects only containing
other strains should be deleted. The original implementation had a bug
where the initial sets of filtered objects could contain duplicates because
Spectrum objects would get added to the list of spectra to retain multiple
times if they had multiple strains in the "include" list. This commit
fixes that by using a set instead of a list.