Some updates to the webapp scoring output
The latest implementation of Metcalf scoring may return a mix of Spectrum
and MolFam results when given GCF inputs, but some parts of the webapp
code still assumed the results would only be Spectrum objects. This updates
a couple of the methods that generate the HTML output to account for this,
mainly by extracting the lists of Spectrum objects from MolFams where
they are encountered and appending them to the existing list of spectra.
This allows the rest of the code to remain mostly as-is.

Also a couple of tweaks to the object counts displayed so that they make
more sense in these circumstances.