Update Rosetta scoring to handle hybrid BGCs
The previous version of the Rosetta code hadn't been updated to track the
API change that allows BGCs to potentially have multiple parents. This
updates the get_links method so that when a hybrid BGC is encountered in
the results and bgc_to_gcf mode is active, an entry will be created for
each of the GCFs in the bgc.parents set. If bgc_to_gcf mode is inactive,
it will still create a single entry for the BGC itself as before.