Adding a workaround for occasional missing MiBIG BGC JSON files
See https://github.com/sdrogers/nplinker/issues/60#issuecomment-1086722952
and subsequent comments. This commit adds a few new methods to the genomics
and downloader modules so that if a missing MiBIG BGC is found during the
loading process, an automatic attempt to retrieve the JSON data can be made
without any action by the user. The downloaded file is placed in the
existing mibig_json folder and a new MiBIGBGC object is created from it,
after which the loading process can proceed as normal.