Remove class `SingletonFamily` and MF attribute `id`
Singleton molecular families can be well represented with the class `MolecularFamily`, and it's not needed to keep a specific class `SingletonFamily` for them, so this PR removed it.

As for the family id, the class `SingleFamily` automatically set it (i.e.`"singleton-" + uuid`), while the class `MolecularFamily` requires users to provide the family id during initiation. It makes more sense to let users or loaders of molecular family to decide how to set family id. So in the `GNPSMolecularFamilyLoader` class, the name of singleton MF was updated to `"singleton-" + spectrum_id`.

The attribute `id` of MolecularFamily is unused by anywhere, so it's removed too.