update logcis of loading mibig BGCs and strains
Loading mibig data is optional, controlled by the setting `self._use_mibig`. Mibig data  is genomics info, so moving the loading of mibig to `self._load_genomics`.

Usually mibig has >2k bgcs, most of which will not be used in nplinker. So it's better to add the used mibig bgcs only to the attribute `self.bgcs` , and add the used mibig strains only to attribute `self.strains`.