remove loading of optional data
The optional data, including GNPS params.xml file and description text, are not needed for core business of NPLinker. To keep the loading process simple (to keep refactored NPLinker as a minimum viable product), the loading of optional data is removed.  If these data are needed in the future, specific loaders should be added for them.