Redesign the template of nplinker config file
This PR redesigned the template of nplinker config file `nplinker.toml`.

The new design does not allow users to specify the path to their data files, instead they should always put data files to default paths defined by NPLinker (the default paths will be provided in the next PR).

To keep it simple, some settings related to code that has not been refactored are not included in the new design. They can be added later if really needed.


### Major changes:
- redesign the template of config file `nplinker.toml`.
- update default config values in `nplinker_default.toml`
- add validations for all configs/settings
- update unit tests for new configs/settings

Note: This PR only updated the template config file, but the usage of the configs in relevant code is not updated. Next PR will update the usage. So it's OK to see the failed test tests on `tests/scorings` and `tests/test_nplinker_local.py` .