add class DatasetArranger
This is a big PR to implement the pipelines of data arranging, which enables the local and podp modes.

Arranging data means
- creating data folders in the `root_dir`
-  downloading dataset if needed (e.g. for podp mode)
-  validating dataset downloaded or provided by users

Basically, it means all steps needed to make data ready for loading. 

The pipelines of arranging data for different types of data are displayed in the diagram of #117.

To keep the data arranging workflow simple, we use fixed project directory structure (see #163) with fixed dir and file names (see `globals.py`).

To use nplinker, users are required to
- create a `root_dir` manually and use it as the root directory of the nplinker project
- provide a config file `nplinker.toml` and put it in the `root_dir` 

**Major changes**
- Added file `arranger.py` including the class `DatasetArranger ` and some validation functions, which implement the pipelines of arranging data

- Clean/remove/update some files to make the arrangers work (some may need further refactoring in future PRs)
    -  cleaned `runbigscape.py`
    - Deleted `downloader.py` and its tests, which is replaced by `DatasetArranger`
    - Updated `loader.py` and `nplinker.py` to use the `DatasetArranger`

- Added integration tests for the arranger (tests passed)
  - Created `nplinker_local_mode.toml`
  - Updated `tests/conftest.py` 
   - Updated `test_nplinker_local.py` to test the `local mode` 
 

Tests on podp mode also passed on my local machine. Due to the cost of running bigscape, the tests will be added to the codebase in next PRs.