Fix github actions for large file storage
We used large file storage (lfs) for handling large test dataset (zip files). 

The github action " [actions/checkout]( https://github.com/actions/checkout)" does not use lfs automatically, leading to failed tests, and so we have to set it explicitly (see the [usage](https://github.com/actions/checkout)).

However, it's not free to use lfs and there is storage and bandwidth limit (see [the comment](https://github.com/NPLinker/nplinker/pull/224#issuecomment-1996709712)), which we do not want.

This PR stopped the use of lfs and implemented an alternative way to handle large files. The approach is to put the large files in Zenodo, and download it before running tests if there is no local cache.

Major changes:
- Stopped the use of lfs ([instructions](https://gist.github.com/everttrollip/198ed9a09bba45d2663ccac99e662201))
- Separated tests to unit tests and integration tests
- Uploaded large test dataset to Zenodo
- Implemented code to prepare test datasets and folders before running tests
- Excluded whole tests folder for distribution (i.e. pypi package)